Fixed all warnings in backend
