+++ backend/src/app/v2/order.rs
@@ -1,14 +1,15 @@
 use std::collections::HashMap;
 use serde::{Deserialize, Serialize};
 use actix_web::{web, HttpResponse};
-use futures::{FutureExt, TryFutureExt, compat::Future01CompatExt, SinkExt};
+use futures::{FutureExt, TryFutureExt, compat::Future01CompatExt};
 use crate::app::AppState;
-use crate::app::v2::stock::{ReadStock, StockInResponse, StockToRemove, RemoveStock};
+use crate::app::v2::stock::{ReadStock, StockInResponse};
 use crate::http::v2::manufacturing::{RecipeRequest, RecipeResponse, SendPartsRequest};
 use crate::app::v1::items::{ReadItems, ItemResponse};
 use crate::http::v2::accounting::{ExpenseRequest, ExpenseResponse};
 use crate::app::v2::items::{ReceiveItemsRequest, ItemInRequest};
 use crate::app::v2::returns::{ProductRequest, PartRequest};
+use crate::db::v2::stock::{RemoveStock, StockToRemove};
 
 #[derive(Debug, Deserialize)]
 pub struct OrderRequest {
@@ -92,7 +93,6 @@ pub async fn manufacturing_order(
     order: OrderRequest,
     stock: HashMap<String, StockInResponse>,
 ) -> Result<(), ()> {
-    let db = &state.db;
     let http = &state.http;
 
     // MANUFACTURING FLOW //////////////////////////////////////////////////////
@@ -161,7 +161,7 @@ pub async fn manufacturing_order(
                 parts_to_order.insert(part, needed_quantity - stock_quantity.quantity);
             }
             // If we have enough stock, we don't need to order any of this part
-            Some(stock_quantity) => {
+            Some(_) => {
                 debug!("We have all {} of {} that we need!", needed_quantity, &part);
                 parts_to_order.insert(part, 0);
             },
